+++ projects/modular_llm/finetune_experts.py
@@ -160,7 +160,7 @@ def finetune_with_nevergrad(args: FinetuneConfig, dm):
         wandb.config.update(args)
 
     from mttl.evaluators.rouge_evaluator import RougeEvaluator
-    from projects.modular_llm.src.nevergrad_opt import NGRoutingOptimizer
+    from mttl.models.nevergrad_opt import NGRoutingOptimizer
 
     library = retrieve(args, args.finetune_task_name, args.sk, retrieve_with="random")
     assert (
@@ -206,7 +206,7 @@ def finetune_with_nevergrad(args: FinetuneConfig, dm):
         wandb.config.update(args)
 
     from mttl.evaluators.rouge_evaluator import RougeEvaluator
-    from projects.modular_llm.src.nevergrad_opt import NGRoutingOptimizer
+    from mttl.models.nevergrad_opt import NGRoutingOptimizer
 
     lib_location = f"/tmp/{args.library_id}"
     os.makedirs(lib_location, exist_ok=True)
